requirements.txt
httpx==0.27.2
httpx-sse==0.4.0
streamlit==1.37.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
openai==1.43.0
